fix(planner): Suppress unused variable warnings in tests
Add [[maybe_unused]] to test variables that exist to build e-graph
structure but aren't directly referenced in assertions.

Change test fixture's leaf() disambiguator parameter from int to
uint64_t to match the underlying ENode type, allowing fuzzer crash
reproducers to be copied into tests without truncation warnings.

+++ src/planner/test/test_egraph_fixture.hpp
@@ -47,7 +47,8 @@ class EGraphTestBase : public ::testing::Test {
   // ---------------------------------------------------------------------------
 
   /// Create a leaf node with the given op and optional disambiguator.
-  auto leaf(Op op, int disambiguator = 0) -> EClassId { return egraph.emplace(op, disambiguator).eclass_id; }
+  /// Note: disambiguator is uint64_t to match ENode and allow fuzzer crash reproducers.
+  auto leaf(Op op, uint64_t disambiguator = 0) -> EClassId { return egraph.emplace(op, disambiguator).eclass_id; }
 
   /// Create a node with the given op and children.
   template <typename... Children>
